+++ Robust.UnitTesting/Shared/Map/GridFixtures_Tests.cs
@@ -14,6 +14,7 @@ namespace Robust.UnitTesting.Shared.Map
     /// <summary>
     /// Tests whether grid fixtures are being generated correctly.
     /// </summary>
+    [Parallelizable(ParallelScope.All)]
     [TestFixture]
     public sealed class GridFixtures_Tests : RobustIntegrationTest
     {